修正技能动画播放bug 框架更新

+++ testdemo/src/com/blackvine/manager/SingleAssetManager.java
@@ -1,11 +1,10 @@
 package com.blackvine.manager;
 
+
 import java.io.File;
 
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.blackvine.constant.Constants;
 import com.blackvine.tools.MyFileNameFilter;
 
@@ -42,20 +41,25 @@ public class SingleAssetManager extends AssetManager {
 	public void addTextureAsset() {
 		// skillFile = new File("assets" + File.separator + "skill"
 		// + File.separator);
+		
 		skillFile = new File(Constants.SKILL_DIR);
 		skillsDir = skillFile.listFiles();
+		System.out.println(skillsDir.length);
 		filter = new MyFileNameFilter();
-		for (int i = 0; i < skillsDir.length; i++) {
-			// filePath = filePath.append("skill/" + i + "/");
-			filePath = filePath.append(Constants.SKILL_DIR + i);
-			singleSkill = new File(filePath.toString());
-			singleSkillDir = singleSkill.listFiles(filter);
-			for (int j = 0; j < singleSkillDir.length; j++) {
-				this.load(Constants.SKILL_DIR + i + "/" + j + ".png",
-						Texture.class);
-			}
-			if (filePath.length() > 0) {
-				filePath = filePath.delete(0, filePath.length());
+		if (skillsDir != null) {
+			for (int i = 0; i < skillsDir.length; i++) {
+				// filePath = filePath.append("skill/" + i + "/");
+				filePath = filePath.append(Constants.SKILL_DIR + i);
+				singleSkill = new File(filePath.toString());
+				singleSkillDir = singleSkill.listFiles(filter);
+				for (int j = 0; j < singleSkillDir.length; j++) {
+//					this.load(Constants.SKILL_DIR + i + "/" + j + ".png",
+//							Texture.class);
+					this.load("skill/" + i + "/" + j + ".png", Texture.class);
+				}
+				if (filePath.length() > 0) {
+					filePath = filePath.delete(0, filePath.length());
+				}
 			}
 		}
 	}
